Finished adding Javadocs to all of the command classes.  Also started to flesh out the ShooterModeShootWithPose command.  The command should now try and move the swerve modules into the brake position, and calculate the distance between the robot and the blue speaker.  It uses two new interpolation tables, kShooterSpeedTable and kShooterArmTable, but the current values are just for debugging.

+++ src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -31,7 +31,7 @@ public class ShooterSubsystem extends SubsystemBase {
   //Make variables
   public String m_shooterMode = "";
   public boolean m_readyToShoot = false;
-  public int m_desiredPosition = 0;
+  public double m_desiredPosition = 0;
   public double m_desiredSpeed = 0;
 
   /*
@@ -177,7 +177,7 @@ public class ShooterSubsystem extends SubsystemBase {
    * Sets the position of the shooter arm.
    * @param position The position of the shooter arm, in rotations.
    */
-  public void setShooterArmPosition(int position) {
+  public void setShooterArmPosition(double position) {
     System.out.println("ShooterSubsystem: setShooterArmPosition");
     //m_ShooterArmRotateLeft.setControl(ShooterArmPosition.withPosition(position));
     m_desiredPosition = position;
